Add files via upload
This D64 file can be uploaded into the VICE C64 emulator. It contains the multicolor, modified version of Catacombs currently. It is not complete yet, but I figured you like to take a look. The file is called CATACOMBS1.5. There are other copies there and files to view for your enjoyment.